backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis
redis==5.0.1

# Azure
azure-storage-blob==12.19.0
openai==1.9.0

# Payments
stripe==7.12.0

# Notifications
sendgrid==6.11.0

# Utilities
httpx==0.26.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.0
isort==5.13.2
